Refactor FastAPI application structure and add configuration management
- Introduced a new configuration management system using Pydantic settings.
- Restructured the FastAPI application by moving models and authentication logic to separate modules.
- Added system information and health check endpoints with proper response models.
- Updated .gitignore to include IDE-specific files.
- Set up logging configuration for better monitoring and debugging.
