+++ .gitignore
@@ -173,3 +173,6 @@ cython_debug/
 # PyPI configuration file
 .pypirc
 .qodo
+
+# IDE
+.idea
