+++ app/conf/logging.py
@@ -1,0 +1,19 @@
+import os
+
+
+LOG_LEVEL = os.getenv("LOG_LEVEL", default="DEBUG")
+LOGGING_CONFIG = {
+    "version": 1,
+    "disable_existing_loggers": False,
+    "formatters": {
+        "standard": {
+            "format": "[%(asctime)s] %(levelname)s [%(filename)s:%(lineno)s] %(message)s",
+            "datefmt": "%d.%m.%Y %H:%M:%S",
+        },
+    },
+    "handlers": {"console": {"class": "logging.StreamHandler", "formatter": "standard"}},
+    "loggers": {
+        "app": {"handlers": ["console"], "level": LOG_LEVEL, "propagate": False},
+        "fastapi": {"handlers": ["console"], "level": LOG_LEVEL, "propagate": False},
+    },
+}
