+++ app/main.py
@@ -1,77 +1,20 @@
-import platform
-from datetime import datetime
+import uvicorn
+from fastapi import FastAPI
 
-from pydantic import BaseModel
-from fastapi import FastAPI, HTTPException, Depends
-from pydantic_settings import BaseSettings
-
-
-class Settings(BaseSettings):
-    """Application settings loaded from environment variables."""
-
-    API_TOKEN: str
-    SERVICE_TOKENS: dict[str, str]
-    ENABLE_SWAGGER: bool = False
-
-    class Config:
-        env_file = ".env"
+from app.conf.app import settings
+from app.routers import system_router
 
 
-class SystemInfo(BaseModel):
-    """Response model for system information endpoint."""
-
-    current_time: datetime
-    os_version: str
-
-
-class HealthCheck(BaseModel):
-    """Response model for health check endpoint."""
-
-    status: str
-    timestamp: datetime
-
-
-settings = Settings()
+# TODO: setup logging config
 app = FastAPI(
     title="System Info API",
     description="API for retrieving system information",
     docs_url="/docs/" if settings.ENABLE_SWAGGER else None,
-    redoc_url="/redoc" if settings.ENABLE_SWAGGER else None,
+    redoc_url="/redoc/" if settings.ENABLE_SWAGGER else None,
 )
+app.include_router(system_router)
 
+if __name__ == "__main__":
+    import uvicorn
 
-def verify_token(token: str) -> bool:
-    """Verify the API token."""
-    return token == settings.API_TOKEN
-
-
-async def get_token(
-    authorization: str = Depends(lambda x: x.headers.get("Authorization", "")),
-) -> str:
-    """Dependency for token verification."""
-    if not authorization.startswith("Bearer "):
-        raise HTTPException(status_code=401, detail="Invalid authorization header")
-
-    token = authorization.split(" ")[1]
-    if not verify_token(token):
-        raise HTTPException(status_code=401, detail="Invalid token")
-
-    return token
-
-
-@app.get("/api/system-info/", response_model=SystemInfo)
-async def get_system_info(_: str = Depends(get_token)) -> SystemInfo:
-    """Get current system information."""
-    return SystemInfo(
-        current_time=datetime.now(),
-        os_version=platform.platform(),
-    )
-
-
-@app.get("/api/health/", response_model=HealthCheck)
-async def health_check() -> HealthCheck:
-    """Health check endpoint."""
-    return HealthCheck(
-        status="healthy",
-        timestamp=datetime.now(),
-    )
+    uvicorn.run(app, host="0.0.0.0", port=8000)
