+++ app/models.py
@@ -1,0 +1,17 @@
+from datetime import datetime
+
+from pydantic import BaseModel
+
+
+class SystemInfo(BaseModel):
+    """Response model for system information endpoint."""
+
+    current_time: datetime
+    os_version: str
+
+
+class HealthCheck(BaseModel):
+    """Response model for health check endpoint."""
+
+    status: str
+    timestamp: datetime
