+++ app/routers/__init__.py
@@ -1,0 +1,3 @@
+from .system import router as system_router
+
+__all__ = ["system_router"]
